Drop the carbonate requirement, pull in whisper-fill.py
whisper-fill.py has been accepted upstream into Whisper by the Graphite
project.  So I imagine it may not stay in Carbonate forever.  Also, I'm
not using Carbonate further in my cluster work.  So let's drop the
dependency.

The Graphite project pulled this into Whisper, but made it a script file
instead of part of the library.  So, I've just pulled in the entire
script file and integrated it into this module.  Rather than doing
weird tricks to find the script and import it.