Require python-swiftclient that supports timeouts
We now set a 60 second timeout on HTTP operations for Swift.
Previously, and with older versions of python-swiftclient there was no
timeout.  A hung connection that the server forgot about would hang
forever and not allow the backup process to complete.  Or even log an
error about the missed metric.

This should log an exception when a connection is hung / timeout that we
can at least dig out of the log files.  It should also successfully
complete.